Add files via upload
i did it! wget seems to not work inside a definition file so had to download the miniconda.sh file to the place where the container would be built, then added the reference code from where they run the .sh file. container finally built and running a shell inside of the container verified that it was running conda 4.11.0